+++ changelog.md
@@ -4,6 +4,8 @@
 
 * Package `args` renamed to `options` (_incompatible changes_)
 * `[fmtutil/table]` Added new package for rendering data as a table
+* `[usage]` Improved output about a newer version
+* `[usage]` Increased code coverage (0.0% → 100%)
 * `[usage]` Code refactoring
 
 ---